// apps/web/services/rankings/types.ts
export type ResultRankingRow = {
  result_id: number;
  attempt_number: number | null;
  result_value: number;
  rank: number;
  position: number;
  person_id: string;
  person_name: string;
  country_id: string;
  country_name: string;
  country_iso2: string;
  continent_id: string;
  competition_id: string;
  competition_name: string;
  record_code: string;
  total_count?: number;
};